requirements.txt
urllib3
aiohttp
python-dotenv